profile-chatbot-api/requirements.txt
fastapi==0.115.9
uvicorn==0.34.2
openai==1.78.1
python-dotenv==1.1.0
